# docs/openapi-3.1.json
{
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/cards": {
      "get": {
        "operationId": "CardsController_list",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "cards"
        ]
      },
      "post": {
        "operationId": "CardsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateCardDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "cards"
        ]
      }
    },
    "/api/v1/cards/{id}": {
      "get": {
        "operationId": "CardsController_get",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "cards"
        ]
      },
      "patch": {
        "operationId": "CardsController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateCardDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "cards"
        ]
      },
      "delete": {
        "operationId": "CardsController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardPurgeDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "cards"
        ]
      }
    },
    "/api/v1/cards/{id}/reset": {
      "post": {
        "operationId": "CardsController_reset",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CardPurgeDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "cards"
        ]
      }
    },
    "/api/v1/cards/{id}/reconcile": {
      "post": {
        "operationId": "CardsController_reconcile",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReconcileCardDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "cards"
        ]
      }
    },
    "/api/v1/cards/{id}/ledger": {
      "get": {
        "operationId": "CardsController_ledger",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "type",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "OPENING_BALANCE",
                "PURCHASE",
                "INSTALLMENT_PRINCIPAL",
                "INTEREST",
                "FEE",
                "ANNUAL_FEE",
                "PAYMENT",
                "REFUND",
                "ADJUSTMENT",
                "REVERSAL"
              ]
            }
          },
          {
            "name": "from",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "to",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "type": "number"
            }
          },
          {
            "name": "cursor",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "cards"
        ]
      }
    },
    "/api/v1/cards/{id}/statements": {
      "get": {
        "operationId": "CardsController_listStatements",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "type": "object"
                  }
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "cards"
        ]
      }
    },
    "/api/v1/cards/{id}/statements/current": {
      "get": {
        "operationId": "CardsController_currentCycle",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "cards"
        ]
      }
    },
    "/api/v1/cards/{id}/statements/{statementId}": {
      "get": {
        "operationId": "CardsController_getStatement",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "statementId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object"
                }
              }
            }
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "cards"
        ]
      },
      "patch": {
        "operationId": "CardsController_updateStatement",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "statementId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStatementDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "access-token": []
          }
        ],
        "tags": [
          "cards"
        ]
      }
    },
...
  },
  "info": {
    "title": "Cuentas API",
    "description": "API para el control de flujo de efectivo y la recomendacion de tarjetas de credito. Las recomendaciones son estimaciones basadas en la informacion registrada por el usuario y no constituyen asesoria financiera profesional.",
    "version": "0.1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Access token (disponible a partir de la Fase 2)"
      }
    },
    "schemas": {
...
      "CreateCardDto": {
        "type": "object",
        "properties": {
          "alias": {
            "type": "string",
            "maxLength": 60
          },
          "institution": {
            "type": "string",
            "maxLength": 80
          },
          "last4": {
            "type": "string",
            "pattern": "^\\d{4}$"
          },
          "creditLimit": {
            "type": "number",
            "minimum": 1,
            "maximum": 2147483647
          },
          "annualRateBps": {
            "type": "number",
            "minimum": 0,
            "maximum": 10000
          },
          "annualFee": {
            "type": "number",
            "minimum": 0,
            "maximum": 2147483647
          },
          "annualFeeMonth": {
            "type": "number",
            "minimum": 1,
            "maximum": 12
          },
          "cutDay": {
            "type": "number",
            "minimum": 1,
            "maximum": 31
          },
          "dueDateMode": {
            "type": "string",
            "enum": [
              "FIXED_DAY",
              "DAYS_AFTER_CUT"
            ]
          },
          "dueDay": {
            "type": "number",
            "minimum": 1,
            "maximum": 31
          },
          "dueDaysAfterCut": {
            "type": "number",
            "minimum": 1,
            "maximum": 60
          },
          "dueNonBusinessDayRule": {
            "type": "string",
            "enum": [
              "PREVIOUS",
              "NEXT",
              "NONE"
            ]
          },
          "sameDayCutIncluded": {
            "type": "boolean"
          },
          "openingBalance": {
            "type": "number",
            "minimum": 1,
            "maximum": 2147483647
          },
          "openingDate": {
            "type": "string"
          }
        },
        "required": [
          "alias",
          "institution",
          "last4",
          "creditLimit",
          "cutDay"
        ]
      },
      "UpdateCardDto": {
        "type": "object",
        "properties": {
          "alias": {
            "type": "string",
            "maxLength": 60
          },
          "institution": {
            "type": "string",
            "maxLength": 80
          },
          "creditLimit": {
            "type": "number",
            "minimum": 1,
            "maximum": 2147483647
          },
          "annualRateBps": {
            "type": "number",
            "minimum": 0,
            "maximum": 10000
          },
          "annualFee": {
            "type": "number",
            "minimum": 0,
            "maximum": 2147483647
          },
          "annualFeeMonth": {
            "type": "number",
            "minimum": 1,
            "maximum": 12
          },
          "cutDay": {
            "type": "number",
            "minimum": 1,
            "maximum": 31
          },
          "dueDateMode": {
            "type": "string",
            "enum": [
              "FIXED_DAY",
              "DAYS_AFTER_CUT"
            ]
          },
          "dueDay": {
            "type": "number",
            "minimum": 1,
            "maximum": 31
          },
          "dueDaysAfterCut": {
            "type": "number",
            "minimum": 1,
            "maximum": 60
          },
          "dueNonBusinessDayRule": {
            "type": "string",
            "enum": [
              "PREVIOUS",
              "NEXT",
              "NONE"
            ]
          },
          "sameDayCutIncluded": {
            "type": "boolean"
          },
          "status": {
            "type": "string",
            "enum": [
              "ACTIVE",
              "INACTIVE"
            ]
          }
        }
      },
      "CardPurgeDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "maxLength": 300
          }
        },
        "required": [
          "reason"
        ]
      },
      "ReconcileCardDto": {
        "type": "object",
        "properties": {
          "reportedBalance": {
            "type": "number",
            "minimum": 0,
            "maximum": 2147483647
          },
          "asOfDate": {
            "type": "string"
          },
          "reason": {
            "type": "string",
            "maxLength": 300
          }
        },
        "required": [
          "reportedBalance",
          "asOfDate",
          "reason"
        ]
      },
      "UpdateStatementDto": {
        "type": "object",
        "properties": {
          "noInterestPaymentReported": {
            "type": "number",
            "minimum": 0,
            "maximum": 2147483647
          },
          "minimumPaymentReported": {
            "type": "number",
            "minimum": 0,
            "maximum": 2147483647
          }
        }
      },
...
    }
  }
}